refactor(core-api-library): Move `DocumentType` to `DocumentManager`
PIA-1298
