+++ core-api-library/library/src/main/java/net/gini/android/core/api/DocumentRemoteSource.kt
@@ -131,14 +131,4 @@ abstract class DocumentRemoteSource(
         const val HEADER_LOCATION_KEY = "location"
     }
 
-    enum class DocumentType(val apiDoctypeHint: String) {
-        BANK_STATEMENT("BankStatement"),
-        CONTRACT("Contract"),
-        INVOICE("Invoice"),
-        RECEIPT("Receipt"),
-        REMINDER("Reminder"),
-        REMITTANCE_SLIP("RemittanceSlip"),
-        TRAVEL_EXPENSE_REPORT("TravelExpenseReport"),
-        OTHER("Other");
-    }
 }
